+++ src/main/java/Duke/Commands/AddCommand.java
@@ -1,3 +1,18 @@
+package Duke.Commands;
+
+import Duke.Exceptions.BackwardsTimeException;
+import Duke.Exceptions.NoDescriptionException;
+import Duke.Exceptions.UnknownCommandException;
+import Duke.Exceptions.UnknownTimeException;
+import Duke.Tasks.Deadline;
+import Duke.Tasks.Event;
+import Duke.Tasks.Task;
+import Duke.Tasks.ToDo;
+import Duke.Tools.Parser;
+import Duke.Tools.Storage;
+import Duke.Tools.TaskList;
+import Duke.Tools.Ui;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
