+++ src/main/java/Duke/Commands/CheckCommand.java
@@ -1,3 +1,9 @@
+package Duke.Commands;
+
+import Duke.Tools.Storage;
+import Duke.Tools.TaskList;
+import Duke.Tools.Ui;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
