+++ src/main/java/Duke/Duke.java
@@ -1,10 +1,13 @@
+package Duke;
+
+import Duke.Commands.Command;
+import Duke.Tools.Parser;
+import Duke.Tools.Storage;
+import Duke.Tools.TaskList;
+import Duke.Tools.Ui;
+
 import java.io.File;
 import java.io.IOException;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeFormatterBuilder;
-import java.time.temporal.ChronoField;
-import java.util.Scanner;
 
 /**
  * Encapsulates the logic of a Chat bot
