+++ src/main/java/Duke/Tools/Parser.java
@@ -1,3 +1,8 @@
+package Duke.Tools;
+
+import Duke.Commands.*;
+import Duke.Exceptions.UnknownCommandException;
+
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeFormatterBuilder;
 import java.time.temporal.ChronoField;
